refactor(markets): Flatten markets file structure
