+++ js/src/pages/markets/add-market/add-market-modal/multilingual-plugin-prompt/index.test.js
