+++ js/src/pages/markets/edit-market-modal/shipping-notice/index.scss
