+++ js/src/pages/markets/market-data-views/index.test.js
