+++ js/src/pages/markets/markets-dashboard/index.js
@@ -11,8 +11,8 @@ import AppNotice from '~/components/app-notice';
 import AppSpinner from '~/components/app-spinner';
 import useDataViewsScript from '~/hooks/useDataViewsScript';
 import useSettings from '~/hooks/useSettings';
-import MarketsHeader from './markets-header';
-import MarketDataViews from './market-data-views';
+import MarketsHeader from '../markets-header';
+import MarketDataViews from '..//market-data-views';
 import './index.scss';
 
 const MarketsDashboard = () => {
