+++ js/src/pages/markets/markets-header/index.scss
