+++ js/src/pages/markets/utils/getShippingRateLabel.js
@@ -13,7 +13,7 @@ import { SHIPPING_RATE_METHOD } from '~/constants';
 import {
 	GOOGLE_MERCHANT_CENTER_URL,
 	WC_SHIPPING_SETTINGS_URL,
-} from './constants';
+} from '../constants';
 
 /**
  * Returns the description shown under the "Markets" heading
@@ -63,24 +63,4 @@ export const getShippingRateLabel = ( shippingRate ) => {
 		default:
 			return null;
 	}
-};
-
-/**
- * Checks for validation errors in the form values.
- *
- * @param {Object} values The form values to check for errors.
- * @param {Array} values.countries The list of selected countries.
- * @return {Object} An object containing error messages for each invalid field.
- */
-export const checkErrors = ( values ) => {
-	const errors = {};
-
-	if ( values.countries.length === 0 ) {
-		errors.countries = __(
-			'Please select at least one country.',
-			'google-listings-and-ads'
-		);
-	}
-
-	return errors;
 };
